**What type of PR is this?**
/kind cleanup

**What this PR does / why we need it**:
In requirements.subtract(), we iterate over all containers of the Pod to calculate new quantity.
Such iteration is repeated at the scope of getPodsToPreempt.

Cache resource requests per Pod per resource name (using a map) and pass it to calls of requirements.subtract() in getPodsToPreempt of preemption

For benchmarking, I made the following change (so that difference in number of ns per operation is more visible):
```
diff --git a/pkg/kubelet/preemption/preemption_test.go b/pkg/kubelet/preemption/preemption_test.go
index 132ffdb659..a2205df37f 100644
--- a/pkg/kubelet/preemption/preemption_test.go
+++ b/pkg/kubelet/preemption/preemption_test.go
@@ -146,7 +146,7 @@ func TestEvictPodsToFreeRequests(t *testing.T) {
 func BenchmarkGetPodsToPreempt(t *testing.B) {
        allPods := getTestPods()
        inputPods := []*v1.Pod{}
-       for i := 0; i < maxPods; i++ {
+       for i := 0; i < 1100; i++ {
                inputPods = append(inputPods, allPods[tinyBurstable])
        }
        for n := 0; n < t.N; n++ {
@@ -156,6 +156,7 @@ func BenchmarkGetPodsToPreempt(t *testing.B) {
                                quantity:     parseCPUToInt64("110m"),
                        }}))
        }
+       t.ReportAllocs()
 }

 func TestGetPodsToPreempt(t *testing.T) {
```
Without patch:
```
BenchmarkGetPodsToPreempt-12    	   20000	     65920 ns/op	     306 B/op	      11 allocs/op
PASS
ok  	k8s.io/kubernetes/pkg/kubelet/preemption	2.196s
```
With patch:
```
BenchmarkGetPodsToPreempt-12    	   20000	     63105 ns/op	     306 B/op	      11 allocs/op
PASS
ok  	k8s.io/kubernetes/pkg/kubelet/preemption	2.071s
```
**Which issue(s) this PR fixes**:
Fixes #77697

```release-note
NONE
```