@tedyu 2019-05-10
cc @tallclair @Random-Liu 
This PR reduces repetitive resource.GetResourceRequest calls.