@dashpole 2019-05-17
@tedyu i'm generally not in favor of adding code complexity for performance improvements we can't quantify, and that aren't causing any issues.  I can't seem to find it, but we used to have guidelines around performance-related PRs, which stated this more elegantly...