@tedyu 2019-05-17
@dashpole 
I modified description to include results from benchmark.